backend/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
camelot-py==0.11.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.4
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
ghostscript==0.7
h11==0.14.0
huggingface-hub==0.22.2
idna==3.6
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.2
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==5.1.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.5
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdfminer.six==20231228
pexpect==4.9.0
pi==0.1.2
pikepdf==9.0.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.6.0
pycparser==2.22
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
PySimpleGUI==5.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.21.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
threadpoolctl==3.3.0
tk==0.1.0
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.39.3
trio==0.25.1
trio-websocket==0.11.1
twilio==9.0.2
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
